% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tutorial2_DEG_snRNAseq.R
\docType{data}
\name{Tutorial2_DEG_snRNAseq}
\alias{Tutorial2_DEG_snRNAseq}
\alias{neg_ctrl,}
\alias{pb,}
\alias{posctrl}
\title{Tutorial2_DEG_snRNAseq}
\description{
this dataset contains the psctrl data.frame including
positive controls, neg_ntrl which is a dataframe of negative controls and
pb which is a SingleCellExperiment of data in pseudo-bulk format
}
\references{
\url{data_blah.com}
}
\author{
My Name \email{blahblah@roxygen.org}
}
\keyword{data}
